requirements.txt
django
django-tables2
requests
django-filter